restructure project, add --local setting
