+++ lib/funcs.js
@@ -6,10 +6,10 @@ function loadPlayer(path, player) {
     const Player = require.main.require(name);
     return new Player(player);
   }catch(e){
-    if (e.code === 'MODULE_NOT_FOUND' && e.message.indexOf(name) !== -1) {
-      console.warn('Cannot find player "%s".\n  Did you forget to install it?\n', name);
+    if (e.code === 'MODULE_NOT_FOUND' && e.message.indexOf(path) !== -1) {
+      console.warn('Cannot find player "%s".\n  Did you forget to install it?\n', path);
     } else {
-      console.warn('Error during loading "%s" player:\n  %s', name, e.message);
+      console.warn('Error during loading "%s" player:\n  %s', path, e.message);
     }
     return false;
   }
