+++ lib/input.js
@@ -13,16 +13,21 @@ const optionDefinitions = [
     description: 'Display the server version'
   },
   {
+    name: 'local',
+    alias: 'l',
+    description: 'Play games locally executing the players code directly (not recommended)'
+  },
+  {
     name: 'a',
     alias: 'a',
     typeLabel: '[underline]{file}',
-    description: 'Client 1 for the algorithm competition'
+    description: 'Client 1 for the algorithm competition (for local games only)'
   },
   {
     name: 'b',
     alias: 'b',
     typeLabel: '[underline]{file}',
-    description: 'Client 2 for the algorithm competition'
+    description: 'Client 2 for the algorithm competition (for local games only)'
   },
   {
     name: 'games',
@@ -40,7 +45,7 @@ const optionDefinitions = [
 const sections = [
   {
     header: 'uttt',
-    content: 'Ultimate Tic Tac Toe - Game Server. If you prefer running it as a server use `$ npm start` instead.'
+    content: 'Ultimate Tic Tac Toe - Game Server'
   },
   {
     header: 'Options',
@@ -49,7 +54,7 @@ const sections = [
   {
     header: 'Synopsis',
     content: [
-      '$ uttt [bold]{-a} [underline]{path/to/programOne} [bold]{-b} [underline]{path/to/programTwo}',
+      '$ uttt --local [bold]{-a} [underline]{path/to/programOne} [bold]{-b} [underline]{path/to/programTwo}',
       '$ uttt [bold]{--help}'
     ]
   }
